requirements.txt
PyJWT
Flask==1.0.2
pymongo==3.6.1
passlib==1.7.1
werkzeug==0.14.1
